modified:   api/src/main/java/com/revature/controller/OrderController.java 	modified:   api/src/main/java/com/revature/model/Menu.java 	modified:   api/src/main/java/com/revature/model/OrderDetails.java 	modified:   api/src/main/java/com/revature/model/OrderId.java 	modified:   frontend/src/App.jsx 	modified:   frontend/src/Pages/Menu.jsx 	modified:   frontend/src/Util/AuthContext.jsx 	modified:   frontend/src/components/Order.jsx 	modified:   frontend/src/main.jsx

+++ api/src/main/java/com/revature/model/Menu.java
@@ -1,13 +1,21 @@
 package com.revature.model;
 
-import jakarta.persistence.*;
+import java.math.BigDecimal;
+import java.math.BigInteger;
+import java.time.LocalDateTime;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.PreUpdate;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "menu")
@@ -20,7 +28,7 @@ public class Menu {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "menu_id")
-    private Long menuId;
+    private BigInteger menuId;
 
     @Column(name = "name", nullable = false, unique = true)
     private String name;
